requirements.txt
# Core Simulation
metadrive==0.2.5        # MetaDrive environment (pip install metadrive)
pygame==2.1.2           # Rendering support for MetaDrive

# Math, AI, Planning
numpy>=1.21.0
scipy>=1.7.0

# Path Planning (Hybrid A*)
networkx>=2.6           # For graph utilities if needed
heapq                   # Built into Python, kept for clarity

# Communication
pyyaml>=6.0             # Config file loader (YAML)
cryptography>=41.0.0    # AES encryption for V2V messages
pycryptodome>=3.18.0    # Alternative encryption lib (RSA/AES support)

# Logging & Debugging
matplotlib>=3.4.0       # Visualization of paths and obstacles
pandas>=1.3.0           # Storing logs, evaluation metrics

# Utils
tqdm>=4.62.0            # Progress bars
